Major changes in new experiment modal:
- New experiment name and task selection steps merged into a unique SetNameAndTaskStep.
- Dataset selection now only shows compatible datasets with respect to the selected task.
- Standarized modal styles.
- Move stepper to modal header.
- Responsive task selection.
